Updated and completed labs

+++ Multiple_Classes_Lab/src/main/java/Artwork.java
@@ -1,10 +1,10 @@
 public class Artwork {
     private String title;
     private String artist;
-    private int price;
+    private double price;
     private int nft;
 
-    public Artwork(String title, String artist, int price, int nft) {
+    public Artwork(String title, String artist, double price, int nft) {
         this.title = title;
         this.artist = artist;
         this.price = price;
@@ -19,7 +19,7 @@ public class Artwork {
         return artist;
     }
 
-    public int getPrice() {
+    public double getPrice() {
         return price;
     }
